# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API documentation ikkino",
        "version": "0.1.1"
    },
    "paths": {
...
        "/users-plan": {
            "get": {
                "tags": [
                    "PlansUser"
                ],
                "summary": "Get a list of user plans",
                "operationId": "ace2d7ceb4b8e937adf59cfdc1a0b16a",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PlansUser"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "PlansUser"
                ],
                "summary": "Create a new user plan",
                "operationId": "726f33c709e9fde43e962fae1645c5dd",
                "requestBody": {
                    "description": "Data to create a new user plan",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plan_id",
                                    "user_id",
                                    "date_begin",
                                    "date_end"
                                ],
                                "properties": {
                                    "plan_id": {
                                        "description": "ID плана",
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "description": "ID пользователя",
                                        "type": "integer"
                                    },
                                    "date_begin": {
                                        "description": "Дата начала",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "date_end": {
                                        "description": "Дата окончания",
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlansUser"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid data provided"
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/users-plan/{id}": {
            "get": {
                "tags": [
                    "PlansUser"
                ],
                "summary": "Get a specific user plan",
                "operationId": "72be375c71a13133de87e874cac3ad60",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID плана пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the user plan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlansUser"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plan not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "PlansUser"
                ],
                "summary": "Update a user plan",
                "operationId": "33a8015d1581f048ef0cf19e53094e3f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID плана пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data for updating the user plan",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "plan_id": {
                                        "description": "ID плана",
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "description": "ID пользователя",
                                        "type": "integer"
                                    },
                                    "date_begin": {
                                        "description": "Дата начала",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "date_end": {
                                        "description": "Дата окончания",
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully updated the user plan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlansUser"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plan not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "PlansUser"
                ],
                "summary": "Delete a user plan",
                "operationId": "a7e54c4d66ae517338762cb53fc17e29",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID плана пользователя для удаления",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plan deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Plan deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plan not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "PlansUser": {
                "properties": {
                    "user_id": {
                        "description": "ID пользователя",
                        "type": "integer"
                    },
                    "plan_id": {
                        "description": "ID плана",
                        "type": "integer"
                    },
                    "date_begin": {
                        "description": "Дата начала",
                        "type": "string",
                        "format": "date"
                    },
                    "date_end": {
                        "description": "Дата окончания",
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}